Tests: Add EGLSurfaceTest.MSAAResolveWithEGLConfig8880 test
Test is for the fix:
    Vulkan: Use correct actual FormatID for MSAA swapchain image
    https://crrev.com/c/angle/angle/+/6275968

SM-G996B (Mali-G78) has these errors in the new test without the fix:

VUID-vkCmdResolveImage-srcImage-01386:
    srcImage and dstImage must have been created with the same image
    format

VUID-VkFramebufferCreateInfo-pAttachments-00880:
    If flags does not include VK_FRAMEBUFFER_CREATE_IMAGELESS_BIT, each
    element of pAttachments must have been created with a VkFormat value
    that matches the VkFormat specified by the corresponding
    VkAttachmentDescription in renderPass

It looks like that in the CL with the fix, above device was incorrectly
tested and commit message noted that the fix is cosmetic because RGB8 is
generally emulated with RGBA8. Turns out, Mali-G78 can render into RGB8.

Test: angle_end2end_tests --gtest_filter=EGLSurfaceTest.MSAAResolveWithEGLConfig8880/*
Bug: angleproject:42265147
Change-Id: I7e90dd7f8c9138a439bc77ed5643214ac4260c73
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6375366
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>