[fuchsia] Simplify connecting to service
Use helper function that automatically opens the service root and
connects to the named service inside. This avoids the need to keep a
handle around. This is being done as part of the io2 migration, where
opening the /svc directory requires PERM_READABLE in most cases.

Bug: fuchsia:324111518
Bug: fuchsia:376575307
Change-Id: I3731719f46b7b7cb028e03b936cdf9716f3a8f81
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6388875
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Yuxin Hu <yuxinhu@google.com>