Split SetUniform into inline and noinline path
Makes a single smaller function for the common cases, and setting
uniforms with a stride or boolean conversion is extracted
to separate function calls. This results in the main SetUniform
being about half in assembly size and yields a small but noticeable
performance improvement in my tests on an Android device.
Impact on the binary size is negligible.

Bug: b/383305597
Change-Id: I630ff41cfeea6ce9e24e30cdebfbb58382e67de7
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6422893
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>