Vulkan: Add assert to check only clear is staged on swapchain image
The swap after clear optimizations on swapchain image multisampled
image
https://chromium-review.googlesource.com/c/angle/angle/+/6072416
works with the assumption that the only staged updates on swapchain
multisampled image is clear. This change adds an assert to validate
the assumption holds true before applying the optimization.
If the assert is ever triggered, we will know the assumption is
violated, and we need to look into how to tweak the optimization or
add other optimizations to solve the violated case.

Bug: angleproject:382006939
Change-Id: I7ce23dd7767b7730f1fa3ec89cb6616689339602
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6437217
Reviewed-by: mohan maiya <m.maiya@samsung.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Commit-Queue: Yuxin Hu <yuxinhu@google.com>