Vulkan: Ghost texture's image backing if overwritten
If the texture's image is in use by the GPU but is overwritten
completely, this change releases the old image and creates a fresh one.
If the texture was used in a render pass, this avoids breaking the
render pass.  Otherwise, it allows the new image to be initialized with
VK_EXT_host_image_copy functionality.  In the very least, an unnecessary
barrier is avoided.

As a targeted optimization, this functionality is limited to non-array
2D color textures which are known to benefit from this ghosting in
applications.  Other texture types can be easily added, but need lots of
tests to be added.

Bug: angleproject:42265356
Change-Id: I18a7587599e36f9f70109264ddc1003b24b8b2df
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6456345
Reviewed-by: Yuxin Hu <yuxinhu@google.com>
Auto-Submit: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Cody Northrop <cnorthrop@google.com>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>