Remove part of releaseTexImageInternal in glGenerateMipmap
In glGenerateMipmap, actually clearing the ImageDesc and
releasing texImage in releaseTexImageInternal is not needed.
The following implementation currently is skipped because getImageDesc
would return empty ImageInfo. Remove those part and put the part
of disconnecting the pbuffer from the texture after generating the
mipmap. And end2end test is also added to verify the implementation.

Bug: angleproject:410996562
Change-Id: I0d8ccd628b98aa5063df2688e9899319c32fd97d
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6461524
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>