Vulkan: Do not skip VUID-vkQueueSubmit-pSignalSemaphores-00067
... when "VK_EXT_swapchain_maintenance1" is enabled.

Original problem is a VVL bug, introduced in:
    62e9314e2eb9c58af42f6f1b613d200d5362598d
    layers: Track semaphore is in use by swapchain

It was later partially fixed in:
    88934ae6acf0c8b8ae169df9607e779ac4a70239
    layers: Fix waiting on present fence did not have effect

Bug is only fixed when using present fences (when
"VK_EXT_swapchain_maintenance1" is enabled).

Without present fences, VUID error still prints the following:
  Swapchain image 0 was presented but not re-acquired, so VkSemaphore
  0x41c700000041c7 may still be in use and cannot be safely reused with
  image index 0.
  a) Use a separate semaphore per swapchain image. Index these
     semaphores using the index of the acquired image.

Since problem only happens in a shared present mode, re-acquiring the
image causing "VUID-vkAcquireNextImageKHR-surface-07783". It is also
unclear from the Vulkan spec if the image may be re-acquired in shared
present mode, but it is currently not necessary.

When not using present fences, ANGLE tracks present semaphores by
attaching submit serial to the previous present operation when same
image is presented. This guarantees to use more than one present
semaphore (in a shared present mode) and not to reuse semaphore with
signal operation pending. This is more strict synchronization than what
is suggested by the VUID error in the item (a).

However, it is debatable if this is ANGLE's or VVL error. If it is not
a VVL error, it at least should provide correct message of how to fix
the problem in case of a shared present mode. Theoretically, ANGLE may
fix this error by only using present semaphore for the initial present,
and skip it for all subsequent present operations, since image is
already shared with the presentation engine and present operation is
only used for notification purposes.

Bug: angleproject:408190758
Change-Id: Id0dc3c0efcba72e844ba6b2c241d0076f70fcd51
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6480175
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>