Tests: Use eglGetPlatformDisplay()
From the EGL 1.5 spec:

  Appendix F
  Version 1.5

  EGL version 1.5 was voted out of the Khronos Technical Working Group
  on January 31, 2014, and formally approved by the Khronos Board of
  Promoters on March 14, 2014.
  EGL 1.5 is the sixth release of EGL. It introduces the following new
  features (the EGL extension(s) each feature is based on are also shown
  parenthetically):
  * Platform support:
    – Providing a mechanism for support of multiple platforms (such as
    window systems or offscreen rendering frameworks) in a single EGL
    implementation at runtime (EGL_EXT_platform_base).

Many tests use eglGetPlatformDisplayEXT() which is provided by the EGL
extension EGL_EXT_platform_base. With the promotion of the
EGL_EXT_platform_base functions to core EGL in version 1.5 and ANGLE
supporting EGL 1.5 (as of at least 2019), update the calls to use
eglGetPlatformDisplay().

This is in preparation for running the ANGLE end2end tests in Android,
which only exposes the EGL 1.5 functions, and not the
EGL_EXT_platform_base functions.

Bug: b/391967165
Test: angle_end2end_tests
Change-Id: I58109c3afe270f46db952e124ee3f5c11200ca35
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6552257
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Amirali Abdolrashidi <abdolrashidi@google.com>
Commit-Queue: Tim Van Patten <timvp@google.com>