Vulkan: Reduce kMaxEmptySlots for SharedDescriptorSetCacheKey
There is report that addKey is still showing up in simpleperf. This CL
breaks addKeyImpl into three functions so that SharedFramebufferCacheKey
will still have the same behavior. SharedDescriptorSetCacheKey is
changed to track 64 cache key at maximum, and updateEmptySlotBits() is
never called in SharedDescriptorSetCacheKey. This means the shared cache
key tracking is further limited to usage case where a buffer/texture is
only involved in less than 64 descriptorSets. Otherwise we will not
track the remaining DescriptorSets, which means if this buffer is
released, the corresponding descriptorSet will not immediately
destroyed, and we will rely on cache eviction code to take care of DS
growing problem.

Bug: b/384839847
Change-Id: I99abd17966446377babace6d06cc8f380a71c084
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6581492
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Yuxin Hu <yuxinhu@google.com>
Commit-Queue: Charlie Lao <cclao@google.com>