Changes to support sdk35
Set the sdkVersion to 35, which forces edge to edge mode.
Edge-to-edge mode without changes causes the ANGLE Preferences
settings page to render the title bar on top of the content.
The fix is to add an inset listener in the fragment, to adjust
the window's marginTop by the inset amount.

The preferences page that was being used had been deprecated as it used
the old ListView, so I updated it to the new way, using
PreferenceFragmentCompat which uses a RecyclerView.
This fragment change cascaded a few changes with the Activity itself,
and the theme, all of which are updated for this change
to use PreferenceFragmentCompat.

Tested: Built in the Android tree, flashed on a Pixel device.
Tested in portrait and landscape, light and dark modes.

Bug: b/403455598
Change-Id: I012b2ea2f9f0973f3e0a926a93d2c3b31daeccca
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6581364
Reviewed-by: Yuxin Hu <yuxinhu@google.com>
Commit-Queue: Cody Northrop <cnorthrop@google.com>
Reviewed-by: Solti Ho <solti@google.com>
Auto-Submit: Brian Johnson <hibrian@google.com>