util: Give all ConfigParameters default values
ConfigParameters has a mix of 'bool' and 'Optional<bool>', with some
tests setting the Optional<> values to the default value, inadvertently
creating a dependency on the associated extension.

To reduce the dependency on various extensions and allow more tests to
run, update ConfigParameters to only contain raw bools and initialize
them to the default values defined by the associated spec. This allows
EGLWindow to only require the necessary extension when a non-default
value is set by a test.

This also requires updates to a handful of tests which were relying on
the state of extensions due to non-default values.

Bug: b/279980674
Change-Id: I77a7e102559e4393aadcdbf37886ae80cbbb9e29
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6618033
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Commit-Queue: Tim Van Patten <timvp@google.com>
Reviewed-by: Amirali Abdolrashidi <abdolrashidi@google.com>