Fix glCopyImageSubData validation check part 2
For compressed texture, we also need to check if the copy subregion
starting points (offset) are aligned with the compressed block.

This change also moves *FillsEntireMip check to the
ValidateCompressedRegion() function, as it is easier to do a
per-dimension check.

Bug: b/419048313
Bug: angleproject:42264179
Bug: angleproject:42265887
Change-Id: I0f415e575bbc77bdf5178e2b5fc1260c43e7de4b
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6648198
Commit-Queue: Yuxin Hu <yuxinhu@google.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Auto-Submit: Yuxin Hu <yuxinhu@google.com>
Reviewed-by: Alexey Knyazev <lexa.knyazev@gmail.com>