Fix GetTestConfigAPIFromRenderer() for SystemEGL Testing
Add support for GLESDriverType::SystemEGL to
GetTestConfigAPIFromRenderer() when system EGL testing is enabled with
ANGLE_TEST_ENABLE_SYSTEM_EGL. This allows the correct GPUTestConfig
value to be returned, rather than GPUTestConfig::kAPIUnknown.

The side-effect of this is fixing the expectations file handling for
entries with qualifiers like "VULKAN" and allowing those tests to be
skipped, for example.

Bug: b/279980674
Change-Id: Iaf039549b82f5ba7a047e621bb26968df511a75e
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6681013
Commit-Queue: Tim Van Patten <timvp@google.com>
Reviewed-by: Amirali Abdolrashidi <abdolrashidi@google.com>