tests: Skip ProgramBinariesAcrossPlatforms on SystemEGL+Android
ProgramBinariesAcrossPlatforms.CreateAndReloadBinary/* is failing when
testing the system EGL on Android.

The issue is that the test calls eglWindow->initializeGL(), without
checking all of the required extensions are available first.
Specifically, it needs the same type of checking that was added to
ANGLETestBase with https://crrev.com/c/6622112.

Rather than duplicating that work, we should instead find a way to
refactor things so both ProgramBinariesAcrossPlatforms and ANGLETestBase
can share the implementation.

This would also fix the issue that ProgramBinariesAcrossPlatforms cannot
be skipped with an entry in angle_end2end_tests_expectations.txt, also
due to not inheriting from ANGLETestBase. Due to this, the test needs to
be skipped directly within ProgramBinariesAcrossPlatforms::SetUp(),
before mOSWindow->initialize() is called (and fails).

Bug: b/424595860
Bug: b/429026255
Change-Id: I118a912e871a4693ac529376ddf97c60dd77b1a0
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6694919
Reviewed-by: Cody Northrop <cnorthrop@google.com>
Reviewed-by: Yuxin Hu <yuxinhu@google.com>
Commit-Queue: Tim Van Patten <timvp@google.com>