Disable EnsureLoopForwardProgressTest on iOS
The CL[1] that enabled the optimization of simple loops introduced
the EnsureLoopForwardProgressTest. However, this test relies on
SH_ESSL_OUTPUT, which requires angle_enable_essl to be enabled.
Since this is not currently supported on iOS, this CL disables
the test on iOS until a proper solution is found.

[1] https://chromium-review.googlesource.com/c/angle/angle/+/6575051

Bug: b/426055001
Change-Id: If694be82bad4b4724ace5051c9ee38829a367170
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6703073
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Commit-Queue: Sunny Sachanandani <sunnyps@chromium.org>